Fix: Fix act sed replacement string for github issue url in test-1-create-a-branch.yml
